+++ embark/lib/templates/debug.sh
@@ -1,1 +1,0 @@
-iex --name debug@$(hostname -I) --cookie secret